Small model update + entity creation in tests

+++ domain-model/src/main/java/com/ridesharing/domain/model/ride/gps/position/GpsPosition.java
@@ -18,5 +18,5 @@ public class GpsPosition {
 
     private Float longitude;
 
-    private Float speed;
+    private Integer speed;
 }
